% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pfeifer.R
\name{calculate.ls}
\alias{calculate.ls}
\title{Creates a grid type copula according to a given data set and the Pfeifer proposal}
\usage{
calculate.ls(U, k, m)
}
\arguments{
\item{U}{a matrix of size nx2 with the observed values.}

\item{k}{positive integer indicating the number of subintervals for the U2 variable.}

\item{m}{positive integer indicating the number of subintervals for the U1 variable.}
}
\description{
Returns a list with a matrix with the density over the grid,
a matrix with the quantity of data over the grid, the number of subintervals for the U2 variable,
the number of subintervals for the U1 variable.
}
