% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grid.R
\name{calculate.ml}
\alias{calculate.ml}
\title{Estimates the parameters of a grid type copula according to their log-likelihood and a given data set}
\usage{
calculate.ml(U, k, m, D.ini = NULL)
}
\arguments{
\item{U}{a matrix of size nx2 with the observed values.}

\item{k}{positive integer indicating the number of subintervals for the U2 variable.}

\item{m}{positive integer indicating the number of subintervals for the U1 variable.}

\item{D.ini}{a matrix with the initial values for the density copula.package: the name of the package for numerical optimization.}
}
\description{
Returns a list with a matrix with the density over the grid,
a matrix with the quantity of data over the grid, the number of subintervals for the U2 variable,
the number of subintervals for the U1 variable.
}
