% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_grid.R
\name{cdf.grid}
\alias{cdf.grid}
\title{Calculates the distribution function of a grid copula for a single data point}
\usage{
cdf.grid(u, mg, Vm)
}
\arguments{
\item{u}{vector of size 1x2 with the values of the variables.}

\item{mg}{a grid type copula object.}

\item{Vm}{Volume matrix.}
}
\description{
Returns the corresponding distribution function value.
}
