% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_grid.R
\name{count.grid}
\alias{count.grid}
\title{Counts the data according to a specified grid}
\usage{
count.grid(U, k, m)
}
\arguments{
\item{U}{a matrix of size nx2 with the observed values.}

\item{k}{a positive integer indicating the number of subintervals for the U2 variable.}

\item{m}{a positive integer indicating the number of subintervals for the U1 variable.}
}
\description{
Returns a matrix of size kxm indicating the quantity of data.
}
