% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equalities_grid.R
\name{equalities.grid}
\alias{equalities.grid}
\title{Evaluates the equalities for the parameters of a grid type copula}
\usage{
equalities.grid(x, A)
}
\arguments{
\item{x}{a vector with the density values of a grid type copula.}

\item{A}{a matrix considered as a reference with the size of the grid.}
}
\description{
Returns the corresponding value of the equalities.
}
