% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_grid.R
\name{ll.grid}
\alias{ll.grid}
\title{Calculates the log-likelihood of a grid type copula}
\usage{
ll.grid(x, y)
}
\arguments{
\item{x}{a vector with the density values of a grid type copula.}

\item{y}{a vector with the counting values in the grid.}
}
\description{
Returns the log-likelihood of a grid type copula.
}
