% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_multiplication.R
\name{normal.multiplication}
\alias{normal.multiplication}
\title{Calculates the multiplication of normal densities}
\usage{
normal.multiplication(x, y)
}
\arguments{
\item{x}{a vector with the values of the variable x.}

\item{y}{a vector with the values of the variable y.}
}
\description{
Returns a vector with the values of the product of densities.
}
