% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective_grid.R
\name{objective.grid}
\alias{objective.grid}
\title{Evaluates the negative log-likelihood function associated to a grid type copula}
\usage{
objective.grid(x, A)
}
\arguments{
\item{x}{a vector with the density values of a grid type copula.}

\item{A}{a matrix with the counting values in the grid.}
}
\description{
Returns the corresponding value of the least squares function.
}
