% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_grid.R
\name{pdf.grid}
\alias{pdf.grid}
\title{Evaluates the density of a grid copula for a single data point}
\usage{
pdf.grid(u, mg)
}
\arguments{
\item{u}{vector of size 1x2 with the values of the variables.}

\item{mg}{a grid type copula object.}
}
\description{
Returns the corresponding distribution function values.
}
