% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_cond_grid.R
\name{r.cond.grid}
\alias{r.cond.grid}
\title{Generates a conditional random sample from a grid type copula}
\usage{
r.cond.grid(U, mg)
}
\arguments{
\item{U}{a vector of size nx1 with values of the U1 variable.}

\item{mg}{a grid type copula object.}
}
\description{
Returns a vector of size nx1 with the random sample of the U2 variable.
}
