% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_integrand_grid.R
\name{rho.integrand.grid}
\alias{rho.integrand.grid}
\title{Evaluates the integrand function for the Spearman's rho concordance measure of a grid type copula.}
\usage{
rho.integrand.grid(x, y, gc)
}
\arguments{
\item{x}{a vector with values of the U1 variable.}

\item{y}{a vector with values of the U2 variable.}

\item{gc}{a grid type copula object.}
}
\description{
Returns the corresponding value of the integrand function.
}
