% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_integrand_grid.R
\name{tau.integrand.grid}
\alias{tau.integrand.grid}
\title{Evaluates the integrand function for the Kendall's tau concordance measure of a grid type copula.}
\usage{
tau.integrand.grid(x, y, gc)
}
\arguments{
\item{x}{a vector with values of the U1 variable.}

\item{y}{a vector with values of the U2 variable.}

\item{gc}{a grid type copula object.}
}
\description{
Returns the corresponding value of the integrand function.
}
