% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_density.R
\name{validate.density}
\alias{validate.density}
\title{Check if the matrix is a density of a copula}
\usage{
validate.density(D.ini, k, m)
}
\arguments{
\item{D.ini}{a matrix with the initial values for the density copula.package: the name of the package for numerical optimization.}

\item{k}{positive integer indicating the number of subintervals for the U2 variable.}

\item{m}{positive integer indicating the number of subintervals for the U1 variable.}
}
\description{
Return a logical value
}
